chore: continue assembler restructure
