+++ vm/src/core_s.c
