+++ vm/src/memory.c
